Backend: adding application with fake calculator

+++ src/Wayfinder/Domain/Coordinates.cs
@@ -7,10 +7,10 @@ public class Coordinates
     public Point Point { get; }
     public IImmutableList<Vector> Polyline { get; }
     public Point Location { get; }
-    public decimal Offset { get; }
-    public decimal Station { get; }
+    public double Offset { get; }
+    public double Station { get; }
 
-    public Coordinates(Point point, IImmutableList<Vector> polyline, Point location, decimal offset, decimal station)
+    public Coordinates(Point point, IImmutableList<Vector> polyline, Point location, double offset, double station)
     {
         this.Point = point;
         this.Polyline = polyline;
